+++ gamevault/UserControls/CommunityUserControl.xaml.cs
@@ -215,7 +215,7 @@ namespace gamevault.UserControls
         {
             if (!LoginManager.Instance.IsLoggedIn())
             {
-                MainWindowViewModel.Instance.AppBarText = "You are not logged in";
+                MainWindowViewModel.Instance.AppBarText = "You are not logged in or offline";
                 return;
             }
             if (uiBtnReloadUser.IsEnabled == false || (e.GetType() == typeof(KeyEventArgs) && ((KeyEventArgs)e).Key != Key.F5))